test: introduce common class for background task ITs
and use it or archiver job tests
