+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/tasks/archiver/AbstractProcessInstanceArchiverJobIT.java
@@ -64,7 +64,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldArchiveLoneProcessInstance(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -82,7 +82,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the process is no longer in the main index
           verifyMoved(listViewTemplate, client, processInstance, "2020-01-01");
@@ -94,7 +94,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
     final int batchSize = 5;
     config.getHistory().setRolloverBatchSize(batchSize);
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -118,7 +118,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(5);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(5);
 
           // check that the first batch of processes are no longer in the main index
           for (var i = 0; i < batchSize; i++) {
@@ -132,7 +132,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldOnlyArchiveFinishedProcessInstances(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -153,7 +153,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the finished process is no longer in the main index
           verifyMoved(listViewTemplate, client, finishedInstance, "2020-01-01");
@@ -164,7 +164,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldOnlyArchiveProcessInstancesCompletedAfterAWhile(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -185,7 +185,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the finished process is no longer in the main index
           verifyMoved(listViewTemplate, client, finishedInstance, "2020-01-01");
@@ -196,7 +196,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldArchiveProcessInstanceAndDependentChildListViewEntities(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -223,7 +223,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the process is no longer in the main index
           verifyMoved(listViewTemplate, client, processInstance, "2020-01-01");
@@ -238,7 +238,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
     // set smaller reindex batch size to verify things work when multiple batches are needed
     config.getHistory().setReindexBatchSize(10);
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -265,7 +265,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the process is no longer in the main index
           verifyMoved(listViewTemplate, client, processInstance, "2020-01-01");
@@ -278,7 +278,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldOnlyArchiveFinishedProcessInstanceAndDependentChildListViewEntities(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -319,7 +319,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the process is no longer in the main index
           verifyMoved(listViewTemplate, client, finishedInstance, "2020-01-01");
@@ -337,7 +337,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldOnlyArchiveFinishedProcessInstanceAndProcessInstanceDependentEntities(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -375,7 +375,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the process is no longer in the main index
           verifyMoved(listViewTemplate, client, finishedInstance, "2020-01-01");
@@ -397,7 +397,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldArchiveProcessInstanceAndProcessInstanceDependentEntities(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -423,7 +423,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
 
           // check that the process is no longer in the main index
           verifyMoved(listViewTemplate, client, processInstance, "2020-01-01");
@@ -438,7 +438,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldNotDeleteDocumentsWhenReindexDoesNotFullySucceed(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given - a finished process instance together with its dependent records
@@ -471,7 +471,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
           // when
           final var result = job.execute().toCompletableFuture();
           Awaitility.await("until the archive job has finished")
-              .atMost(ARCHIVE_TIMEOUT)
+              .atMost(EXECUTE_TIMEOUT)
               .until(result::isDone);
           client.refresh();
 
@@ -492,7 +492,7 @@ public abstract class AbstractProcessInstanceArchiverJobIT<T extends ProcessInst
   @TestTemplate
   void shouldHaveProcessInstanceDependentEntitiesSpecifiedInTests(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
