+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/tasks/archiver/BatchOperationArchiverJobIT.java
@@ -51,7 +51,7 @@ public class BatchOperationArchiverJobIT extends ArchiverJobIT<BatchOperationArc
   @TestTemplate
   void shouldArchiveBatchOperationAndDependantAuditLog(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given - batch operation with numeric ID and matching audit log entry
@@ -71,7 +71,7 @@ public class BatchOperationArchiverJobIT extends ArchiverJobIT<BatchOperationArc
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
           verifyMoved(batchOpTemplate, client, batchOp, "2020-01-01");
           verifyMoved(auditLogTemplate, client, auditLog, "2020-01-01");
         });
@@ -80,7 +80,7 @@ public class BatchOperationArchiverJobIT extends ArchiverJobIT<BatchOperationArc
   @TestTemplate
   void shouldArchiveBatchOperationWithGuidIdWithoutFailure(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given - legacy 8.8 batch operation with GUID id
@@ -101,7 +101,7 @@ public class BatchOperationArchiverJobIT extends ArchiverJobIT<BatchOperationArc
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
           verifyMoved(batchOpTemplate, client, batchOp, "2020-01-01");
           verifyNotMoved(auditLogTemplate, client, auditLog);
         });
@@ -110,7 +110,7 @@ public class BatchOperationArchiverJobIT extends ArchiverJobIT<BatchOperationArc
   @TestTemplate
   void shouldArchiveMixedBatchOperationsAndOnlyMatchNumericDependants(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given - mix of 8.8 GUID and 8.9 numeric batch operation IDs with same end date
@@ -135,7 +135,7 @@ public class BatchOperationArchiverJobIT extends ArchiverJobIT<BatchOperationArc
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(ARCHIVE_TIMEOUT).isEqualTo(2);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(2);
           verifyMoved(batchOpTemplate, client, numericBatchOp, "2020-01-01");
           verifyMoved(batchOpTemplate, client, guidBatchOp, "2020-01-01");
           verifyMoved(auditLogTemplate, client, matchingAuditLog, "2020-01-01");
