+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/tasks/archiver/UsageMetricTUArchiverJobIT.java
@@ -14,7 +14,6 @@ import io.camunda.exporter.config.ExporterConfiguration;
 import io.camunda.search.test.utils.SearchClientAdapter;
 import io.camunda.webapps.schema.descriptors.template.UsageMetricTUTemplate;
 import io.camunda.webapps.schema.entities.metrics.UsageMetricsTUEntity;
-import java.time.Duration;
 import java.time.OffsetDateTime;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
@@ -43,7 +42,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
   @TestTemplate
   void shouldArchiveOldUsageMetricTU(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given
@@ -58,7 +57,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
           final var archived = job.execute();
 
           // then
-          assertThat(archived).succeedsWithin(Duration.ofSeconds(5L)).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
           verifyMoved(template, client, metric, "2020-01-01");
         });
   }
@@ -66,7 +65,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
   @TestTemplate
   void shouldNotArchiveRecentUsageMetricTU(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given - one old metric and one recent metric
@@ -84,7 +83,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
           final var archived = job.execute();
 
           // then - only old metric should be archived
-          assertThat(archived).succeedsWithin(Duration.ofSeconds(5L)).isEqualTo(1);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
           verifyMoved(template, client, oldMetric, "2020-01-01");
           verifyNotMoved(template, client, recentMetric);
         });
@@ -93,7 +92,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
   @TestTemplate
   void shouldArchiveMultipleUsageMetricTUWithSameEndTime(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given - multiple metrics with the same endTime are batched together
@@ -111,7 +110,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
           final var archived = job.execute();
 
           // then - both should be archived to the same dated index in a single batch
-          assertThat(archived).succeedsWithin(Duration.ofSeconds(5L)).isEqualTo(2);
+          assertThat(archived).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(2);
           verifyMoved(template, client, metric1, "2020-03-01");
           verifyMoved(template, client, metric2, "2020-03-01");
         });
@@ -120,7 +119,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
   @TestTemplate
   void shouldArchiveMetricTUFromDifferentMonthsInSeparateBatches(
       final ExporterConfiguration config, final SearchClientAdapter client) throws Exception {
-    withArchiverJob(
+    withTask(
         config,
         (job, resourceProvider) -> {
           // given - metrics from different months require separate execute calls
@@ -138,7 +137,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
           final var firstBatch = job.execute();
 
           // then
-          assertThat(firstBatch).succeedsWithin(Duration.ofSeconds(5L)).isEqualTo(1);
+          assertThat(firstBatch).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
           verifyMoved(template, client, januaryMetric, "2020-01-01");
           verifyNotMoved(template, client, marchMetric);
 
@@ -147,7 +146,7 @@ public class UsageMetricTUArchiverJobIT extends ArchiverJobIT<UsageMetricTUArchi
           final var secondBatch = job.execute();
 
           // then
-          assertThat(secondBatch).succeedsWithin(Duration.ofSeconds(5L)).isEqualTo(1);
+          assertThat(secondBatch).succeedsWithin(EXECUTE_TIMEOUT).isEqualTo(1);
           verifyMoved(template, client, marchMetric, "2020-03-01");
         });
   }
